requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
PyJWT==2.0.0
pyodbc==4.0.30
six==1.15.0
SQLAlchemy==1.3.22
SQLAlchemy-serializer==1.3.4.4
Werkzeug==1.0.1
